style: components/Home에 componets 분리
